refactor: move operations with DTO from MinioService to Controllers
